[✨feat]: RankList 컴포넌트로 분리하여 구현
